feat: Change to library and use a `FindNodeTrait`
Second try after asking:

"Make this a lib and create a FindNodeTrait so we can mock the network and pass
the trait to `Kademlia::new` instead of `send_find_node_request`."

GPT4 suggested using a trait with an async function, but that isn't supported yet in
rust stable. So I've changed the FindNodeTrait::find_node_request to NOT be async.